Rename the track namespace to agv
